Drop all content from GitHub pages
This is already covered by Read the Docs, which is doing the building
and uploading of the docs for us. Since all released versions are
already available on Read the Docs including the latest development
version, there is no need for us to even have the docs for historic
purposes.