Make the index page a redirect to Read the Docs
As users may still visit the webpage due to old links floating around on
the web, it is important to ensure that the old doc page on GitHub pages
redirects to the current one on Read the Docs. The code contained here
tries to perform this redirection through a number of mechanisms. First
it tries to redirect users to Read the Docs with the meta refresh trick
using a negligible timeout. Some browsers don't like no timeout; hence,
why we include a small one. Second if the meta refresh trick fails, we
fallback to JavaScript redirection. Third, if JavaScript redirection
fails either due to an old browser or settings that block it, we resort
to the good old tell the user to redirect themselves with a link page.
That way the users still get the message even if we couldn't help them
on their journey. Finally to ensure this become unnecessary in the long
run, we teach web crawlers that the page to look at is the one on Read
the Docs. That ways search results should point over there instead.